feat: reposition visualization legend
